Fix for issue #2 use the hub proxy password in the update
The proxy password variable was not used to check if it was blank when
the proxy password is reset from a valid value to an empty string which
is valid for non-authenticated proxies.  Update the variable to the
proxy password rather than the password field to reset the username and
password correctly.